pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openmetadata-mcp-agent"
version = "0.1.0"
description = "Conversational governance agent for OpenMetadata. Standalone Python project using data-ai-sdk + LangGraph + OpenAI."
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "Apache-2.0" }
authors = [
    { name = "Guna Palanivel", email = "gunapalanivel@gmail.com" },
    { name = "Priyanka Sen" },
    { name = "Aravind Sai" },
    { name = "Bhawika Kumari" },
]
keywords = [
    "openmetadata",
    "mcp",
    "model-context-protocol",
    "langchain",
    "langgraph",
    "ai-agent",
    "data-governance",
    "metadata",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Official OpenMetadata Python SDK
    "data-ai-sdk==0.1.2",

    # LangGraph orchestration
    "langgraph>=1.0.10,<2.0.0",

    # OpenAI direct (for typed responses + circuit-breaker wrapping)
    "openai>=1.50.0,<2.0.0",

    # FastAPI + ASGI
    "fastapi>=0.110.0,<0.137.0",
    "uvicorn[standard]>=0.27.0,<0.40.0",

    # Pydantic v2
    "pydantic>=2.6.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",

    # Scripts (load_seed, trigger_om_search_reindex) read AI_SDK_TOKEN from repo .env
    "python-dotenv>=1.0.0,<2.0.0",

    # HTTP client (replaces requests; timeout-default)
    "httpx>=0.26.0,<0.30.0",

    # Resilience
    "tenacity>=8.2.0,<10.0.0",
    "pybreaker>=1.0.0,<2.0.0",
    "slowapi>=0.1.9,<0.2.0",

    # Observability
    "structlog>=24.1.0,<26.0.0",
    "prometheus-client>=0.20.0,<0.30.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=9.0.3,<10.0.0",
    "pytest-asyncio>=1.2.0,<2.0.0",
    "pytest-cov>=4.1.0,<7.0.0",
    "respx>=0.20.0,<0.30.0",

    # Linting / formatting
    "ruff>=0.4.0,<1.0.0",

    # Type checking
    "mypy>=1.8.0,<2.0.0",

    # Security scanning
    "bandit[toml]>=1.7.0,<2.0.0",
    "pip-audit>=2.7.0,<3.0.0",

    # Pre-commit
    "pre-commit>=3.6.0,<5.0.0",
]

[project.urls]
Homepage = "https://github.com/GunaPalanivel/openmetadata-mcp-agent"
Repository = "https://github.com/GunaPalanivel/openmetadata-mcp-agent"
Issues = "https://github.com/GunaPalanivel/openmetadata-mcp-agent/issues"
"Hackathon Track" = "https://www.wemakedevs.org/hackathons/openmetadata"

[project.scripts]
copilot = "copilot.api.main:run"

# -----------------------------------------------------------------------------
# Hatchling
# -----------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src/copilot"]

# -----------------------------------------------------------------------------
# Ruff
# -----------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
    "T20",    # flake8-print (no print() in production)
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
    "PT",     # flake8-pytest-style
    "PERF",   # perflint
]
ignore = [
    "E501",   # line-too-long handled by formatter
    "S101",   # assert_used: pytest uses assert
    "S104",   # hardcoded_bind_all_interfaces: SC-1 enforces 127.0.0.1 in code
    "B008",   # function-call-in-default-argument: FastAPI Depends() pattern
]

[tool.ruff.lint.per-file-ignores]
# tests don't need security/print/perf rules (test code runs once, clarity > speed)
"tests/**/*.py" = ["S", "T20", "PERF401"]
"scripts/**/*.py" = ["T20"]      # scripts may print
"src/copilot/models/chat.py" = ["S105"]  # ErrorCode enum string values flagged as "passwords" (false positive)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# -----------------------------------------------------------------------------
# Mypy
# -----------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_unused_ignores = true
disallow_untyped_decorators = true
no_implicit_reexport = true
show_error_codes = true
files = ["src/copilot"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["pybreaker.*", "slowapi.*", "data_ai_sdk.*", "ai_sdk.*"]
ignore_missing_imports = true

# -----------------------------------------------------------------------------
# Pytest
# -----------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
markers = [
    "integration: marks tests as integration (requires running OM container)",
    "e2e: marks tests as end-to-end (requires full stack)",
    "security: marks tests as security-focused",
]

[tool.coverage.run]
source = ["src/copilot"]
branch = true

[tool.coverage.report]
fail_under = 70
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

# -----------------------------------------------------------------------------
# Bandit
# -----------------------------------------------------------------------------
[tool.bandit]
exclude_dirs = ["tests", "ui", ".venv", "venv", "build", "dist"]
skips = ["B101"]   # assert_used in tests
